requirements.txt
# http request and parsing stuff
requests
BeautifulSoup4

# captcha stuff
pillow